[C-826] Move remote-config-instance to saga context (#1713)
