+++ packages/web/src/common/store/service-selection/store.test.js
