+++ packages/web/src/components/service-selection/SelectedServices.tsx
@@ -7,9 +7,13 @@ import { ReactComponent as IconInfo } from 'assets/img/iconInfo.svg'
 import Tooltip from 'components/tooltip/Tooltip'
 import { useSelector } from 'utils/reducer'
 
+import { getSelectedServices } from '../../common/store/service-selection/selectors'
+import {
+  openModal,
+  fetchServices
+} from '../../common/store/service-selection/slice'
+
 import styles from './SelectedServices.module.css'
-import { getSelectedServices } from './store/selectors'
-import { openModal, fetchServices } from './store/slice'
 import { trimServiceName } from './utils'
 
 type ServiceNameProps = {
