+++ packages/web/src/pages/App.js
@@ -47,7 +47,7 @@ import MobileRoute from 'components/routes/MobileRoute'
 import TrendingGenreSelectionPage from 'components/trending-genre-selection/TrendingGenreSelectionPage'
 import AnnouncementPage from 'pages/announcement-page/AnnoucementPage'
 import ArtistDashboardPage from 'pages/artist-dashboard-page/ArtistDashboardPage'
-import AudioRewardsPage from 'pages/audio-rewards-page/AudioRewardsPage'
+import { AudioRewardsPage } from 'pages/audio-rewards-page/AudioRewardsPage'
 import CheckPage from 'pages/check-page/CheckPage'
 import CollectionPage from 'pages/collection-page/CollectionPage'
 import EmptyPage from 'pages/empty-page/EmptyPage'
