+++ packages/web/src/pages/track-page/store/sagas.js
@@ -26,13 +26,13 @@ import {
 } from 'common/store/pages/track/selectors'
 import { getIsReachable } from 'common/store/reachability/selectors'
 import tracksSagas from 'pages/track-page/store/lineups/tracks/sagas'
-import { remoteConfigInstance } from 'services/remote-config/remote-config-instance'
 import { NOT_FOUND_PAGE, trackRemixesPage } from 'utils/route'
 
 export const TRENDING_BADGE_LIMIT = 10
 
 function* watchTrackBadge() {
   const apiClient = yield getContext('apiClient')
+  const remoteConfigInstance = yield getContext('remoteConfigInstance')
   yield takeEvery(trackPageActions.GET_TRACK_RANKS, function* (action) {
     try {
       yield call(waitForBackendSetup)
