+++ packages/web/src/store/sagas.ts
@@ -13,6 +13,7 @@ import explorePageSagas from 'common/store/pages/explore/sagas'
 import reachabilitySagas from 'common/store/reachability/sagas'
 import recoveryEmailSagas from 'common/store/recovery-email/sagas'
 import remoteConfigSagas from 'common/store/remote-config/sagas'
+import serviceSelectionSagas from 'common/store/service-selection/sagas'
 import signOutSagas from 'common/store/sign-out/sagas'
 import artistRecommendationsSagas from 'common/store/ui/artist-recommendations/sagas'
 import deletePlaylistConfirmationModalSagas from 'common/store/ui/delete-playlist-confirmation-modal/sagas'
@@ -28,7 +29,6 @@ import notificationSagas from 'components/notification/store/sagas'
 import passwordResetSagas from 'components/password-reset/store/sagas'
 import remixSettingsModalSagas from 'components/remix-settings-modal/store/sagas'
 import searchBarSagas from 'components/search-bar/store/sagas'
-import serviceSelectionSagas from 'components/service-selection/store/sagas'
 import shareSoundToTikTokModalSagas from 'components/share-sound-to-tiktok-modal/store/sagas'
 import dashboardSagas from 'pages/artist-dashboard-page/store/sagas'
 import rewardsPageSagas from 'pages/audio-rewards-page/store/sagas'
