+++ packages/web/src/store/types.ts
@@ -11,6 +11,7 @@ import { SmartCollectionState } from 'common/store/pages/smart-collection/slice'
 import QueueReducer from 'common/store/queue/slice'
 import { ReachabilityState } from 'common/store/reachability/types'
 import RemoteConfigReducer from 'common/store/remote-config/slice'
+import ServiceSelectionReducer from 'common/store/service-selection/slice'
 import StemsUploadReducer from 'common/store/stems-upload/slice'
 import { CreatePlaylistModalState } from 'common/store/ui/createPlaylistModal/types'
 import { FavoritesPageState } from 'common/store/user-list/favorites/types'
@@ -24,7 +25,6 @@ import MusicConfetti from 'components/music-confetti/store/slice'
 import { PasswordResetState } from 'components/password-reset/store/types'
 import RemixSettingsModalReducer from 'components/remix-settings-modal/store/slice'
 import SearchBarState from 'components/search-bar/store/types'
-import ServiceSelectionReducer from 'components/service-selection/store/slice'
 import { UnfollowConfirmationModalState } from 'components/unfollow-confirmation-modal/store/types'
 import ArtistDashboardState from 'pages/artist-dashboard-page/store/types'
 import { DeactivateAccountState } from 'pages/deactivate-account-page/store/slice'
